% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_footprint_detection.r
\name{AggregateFootprints}
\alias{AggregateFootprints}
\title{Gather equivalent footprints by overlaps (and TF identity) under the same index}
\usage{
AggregateFootprints(footprints.df)
}
\arguments{
\item{footprints.df}{data.frame of footprints as returned by FootprintCharter() or internally by DetectFootprints() or AnnotateFootprints()}
}
\description{
assigns an index to footprints which allows to consider two slightly different footprints as equivalent
given the following condition:
the footprints coordinates overlap by >= 75% of the width of the smaller one and (if given) have identical TF identity
}
