% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_footprint_detection.r
\name{AnnotateFootprints}
\alias{AnnotateFootprints}
\title{Annotate detected TF footprints with user-provided TF motif annotations}
\usage{
AnnotateFootprints(footprints.df, chromosome, TFBSs)
}
\arguments{
\item{footprints.df}{data.frame of footprints as returned by FootprintCharter() or internally by DetectFootprints()}

\item{chromosome}{chromosome of current Region of interest.}

\item{TFBSs}{TF motif annotations. GRanges with at least two metadata columns: TF and absolute.idx for TF identity and motif index, respectively}
}
\description{
Annotate detected TF footprints with user-provided TF motif annotations
}
