% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeWide_wrappers_to_SortingFunctions.R
\name{Arrange_TFBSs_clusters}
\alias{Arrange_TFBSs_clusters}
\title{Convenience function to arrange a list of given TFBSs into clusters}
\usage{
Arrange_TFBSs_clusters(
  TFBSs,
  max_intersite_distance = 75,
  min_intersite_distance = 25,
  max_cluster_size = 10
)
}
\arguments{
\item{TFBSs}{GRanges object of TFBSs}

\item{max_intersite_distance}{maximum allowed distance in base pairs between two TFBS centers for them to be considered part of the same cluster. Defaults to 75.}

\item{min_intersite_distance}{minimum allowed distance in base pairs between two TFBS centers for them not to be discarded as overlapping. Defaults to 25.}

\item{max_cluster_size}{maximum number of TFBSs to be contained in any given cluster. Defaults to 10}
}
\value{
list with two elements: ClusterCoordinates (GRanges object of clusters coordinates) and ClusterComposition (GRangesList of sites for each cluster)
}
\description{
Convenience function to arrange a list of given TFBSs into clusters
}
