% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{cbind.fill.matrix.sparse}
\alias{cbind.fill.matrix.sparse}
\title{Implementation performing a similar operation of rbind.fill.matrix.sparse but for columns}
\usage{
cbind.fill.matrix.sparse(x, y)
}
\arguments{
\item{x}{sparse matrix constructed using the function Matrix::sparseMatrix. Should have Dimnames and dims (e.g. when indexing drop=FALSE)}

\item{y}{sparse matrix constructed using the function Matrix::sparseMatrix. Should have Dimnames and dims (e.g. when indexing drop=FALSE)}
}
\description{
Implementation performing a similar operation of rbind.fill.matrix.sparse but for columns
}
\details{
N.b. only possible fill at the moment is 0
}
