% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{colMeans_drop0}
\alias{colMeans_drop0}
\title{Calculate colMeans after dropping zeros}
\usage{
colMeans_drop0(MethSM)
}
\arguments{
\item{MethSM}{one single molecule sparse matrix}
}
\value{
colMeans (N.b. this is +1 based)
}
\description{
Calculate colMeans after dropping zeros
}
