% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompositePlotting.R
\name{CompositePlot}
\alias{CompositePlot}
\title{Plot composite SMF data}
\usage{
CompositePlot(CompositeDF, span = 0.1, TF)
}
\arguments{
\item{CompositeDF}{the output of the CollectCompositeData function}

\item{span}{the span parameter to pass to geom_smooth}
}
\description{
Will use geom_point with <= 5000 points, geom_hex otherwise
}
