% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs.r
\name{ConversionRate}
\alias{ConversionRate}
\title{Conversion rate}
\usage{
ConversionRate(sampleSheet, genome, chr = 19, clObj = NULL)
}
\arguments{
\item{sampleSheet}{QuasR sample sheet}

\item{genome}{BS genome}

\item{chr}{chromosome to calculate conversion rate on (default: 19)}

\item{clObj}{cluster object to emply for parallel processing created using the parallel::makeCluster function. Defaults to NULL}
}
\value{
Conversion rate
}
\description{
calculate sequencing library conversion rate on a chromosome of choice
}
\examples{
Qinput = paste0(tempdir(), "/NRF1Pair_Qinput.txt")
library(BSgenome.Mmusculus.UCSC.mm10)

if(file.exists(Qinput)){
    # DO NOT RUN
    # clObj = parallel::makeCluster(5)
    # ConversionRatePrecision = ConversionRate(sampleSheet = Qinput, genome = BSgenome.Mmusculus.UCSC.mm10, chr = 19, clObj = clObj)
    # parallel::stopCluster(clObj)
}

}
