% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_calling.r
\name{CoverageFilter}
\alias{CoverageFilter}
\title{Filter Cs for coverage}
\usage{
CoverageFilter(MethGR, thr)
}
\arguments{
\item{MethGR}{Granges obj of average methylation}

\item{thr}{converage threshold}
}
\value{
filtered MethGR
}
\description{
Filter Cs for coverage
}
\examples{

Methylation = qs::qread(system.file(
    "extdata", "Methylation_3.qs", package="SingleMoleculeFootprinting"
    ))
CoverageFilter(MethGR = Methylation[[1]], thr = 20)

}
