% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_calling.r
\name{DetectExperimentType}
\alias{DetectExperimentType}
\title{Detect type of experiment}
\usage{
DetectExperimentType(Samples)
}
\arguments{
\item{Samples}{SampleNames field from QuasR sampleFile}
}
\description{
Detect type of experiment
}
\examples{

CacheDir = ExperimentHub::getExperimentHubOption(arg = "CACHE")
sampleFile = paste0(CacheDir, "/NRF1Pair_sampleFile.txt")
samples = suppressMessages(unique(readr::read_delim(sampleFile, delim = "\t")[[2]]))
DetectExperimentType(samples)

}
