% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{filter_reads_from_MethGR}
\alias{filter_reads_from_MethGR}
\title{Recalculate *_T and *_M values in MethGR object after filtering reads e.g. for conversion rate}
\usage{
filter_reads_from_MethGR(MethGR, MethSM, MethSM_filtered, sampleIndex)
}
\arguments{
\item{MethGR}{GRanges object of methylation call}

\item{MethSM}{Single Molecule methylation matrix}

\item{MethSM_filtered}{Single Molecule methylation matrix after filtering reads}

\item{sampleIndex}{index for sample to treat. It serves as a correspondence between the index of the SM matrix and the order samples appear in the elementMetadata() columns}
}
\value{
MethGR with recalculated counts
}
\description{
Recalculate *_T and *_M values in MethGR object after filtering reads e.g. for conversion rate
}
