% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{FixOverhang}
\alias{FixOverhang}
\title{Fixing overhang before stand collapsing}
\usage{
FixOverhang(MethGR, context, which)
}
\arguments{
\item{MethGR}{Granges obj of average methylation}

\item{context}{context}

\item{which}{"Top"|"Bottom"}
}
\value{
MethGR with fixed overhang
}
\description{
Fixing overhang before stand collapsing
}
