% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_calling_utils.r
\name{full.join.granges}
\alias{full.join.granges}
\title{Utility function to perform the dplyr full_join operation on GRanges object}
\usage{
full.join.granges(MethGR1, MethGR2)
}
\arguments{
\item{MethGR1}{Methylation GRanges as output by the CallContextMethylation() function}

\item{MethGR2}{Methylation GRanges as output by the CallContextMethylation() function}
}
\description{
Utility function to perform the dplyr full_join operation on GRanges object
}
