% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_calling.r
\name{GetQuasRprj}
\alias{GetQuasRprj}
\title{Get QuasRprj}
\usage{
GetQuasRprj(sampleFile, genome)
}
\arguments{
\item{sampleFile}{QuasR pointer file}

\item{genome}{BSgenome}
}
\description{
Get QuasRprj
}
\examples{

library(BSgenome.Mmusculus.UCSC.mm10)
CacheDir <- ExperimentHub::getExperimentHubOption(arg = "CACHE")
sampleFile = paste0(CacheDir, "/NRF1Pair_sampleFile.txt")
QuasRprj = GetQuasRprj(sampleFile, BSgenome.Mmusculus.UCSC.mm10)

}
