% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs.r
\name{GRanges_to_DF}
\alias{GRanges_to_DF}
\title{Manipulate GRanges into data.frame}
\usage{
GRanges_to_DF(GRanges_obj)
}
\arguments{
\item{Single}{GRanges object as returned by CallContextMethylation function}
}
\description{
Inner utility for LowCoverageMethRateDistribution
}
