% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs.r
\name{LowCoverageMethRate_MSE}
\alias{LowCoverageMethRate_MSE}
\title{Low Coverage Methylation Rate MSE}
\usage{
LowCoverageMethRate_MSE(BinnedMethRate)
}
\arguments{
\item{BinnedMethRate}{data.frame as returned by GRanges_to_DF function.}
}
\description{
Calculate Mean squared error (MSE) of methylation rate distribution estimates for low coverage samples
}
