% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs.r
\name{LowCoverageMethRate_RMSE}
\alias{LowCoverageMethRate_RMSE}
\title{Low Coverage Methylation Rate RMSE}
\usage{
LowCoverageMethRate_RMSE(BinnedMethRate)
}
\arguments{
\item{BinnedMethRate}{data.frame as returned by GRanges_to_DF function.}
}
\description{
Calculate Root mean squared error (RMSE) of methylation rate distribution estimates for low coverage samples
}
