% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{MakeBins}
\alias{MakeBins}
\title{Generic function to design bins based on different user case}
\usage{
MakeBins(
  RegionsOfInterest,
  BinType = "singleTF",
  userBin = NULL,
  StrandAware = FALSE
)
}
\arguments{
\item{RegionsOfInterest}{GRanges of length == 1 (for single TFs and promoters) or >= 1 for TF clusters containing the biological loci to sort around}

\item{BinType}{one of "singleTF", "TFcluster", "Promoter". Defaults to "singleTF"}

\item{userBin}{custom user-passed bins in the format list(bin1 = c(start, end), bin2 = c(start, end), ...). Defaults to NULL. If used the arg BinType is ignored}

\item{StrandAware}{TRUE or FALSE (default). If BinType is "singleTF" or "TFcluster", StrandAware argument is ignored}
}
\description{
Generic function to design bins based on different user case
}
