% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_rolling_mean.r
\name{matrix.sliding.window.average}
\alias{matrix.sliding.window.average}
\title{Computes rolling mean}
\usage{
matrix.sliding.window.average(MethSM, window.size = 40, padding = 20)
}
\arguments{
\item{MethSM}{single molecule matrix (dense)}

\item{window.size}{size of the window used to smooth molecules. Defaults to 40}

\item{padding}{padding size. Defaults to 20}
}
\description{
Computes rolling mean
}
