% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{MergeMatrixes}
\alias{MergeMatrixes}
\title{Merged GC and CG matrix (Obsolete since implementation of sparse SM)}
\usage{
MergeMatrixes(matrixes)
}
\arguments{
\item{matrixes}{list of two matrixes to merge, one for GC info and one for CG.}
}
\value{
merged matrix
}
\description{
Merged GC and CG matrix (Obsolete since implementation of sparse SM)
}
\examples{

MergedSM = MergeMatrixes(MethSM)

}
