% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_rolling_mean.r
\name{MethSM.to.dense}
\alias{MethSM.to.dense}
\title{Turn sparse single molecule matrix to dense}
\usage{
MethSM.to.dense(MethSM)
}
\arguments{
\item{MethSM}{sparse MethSM as returned by CallContextMethylation()}
}
\description{
Turn sparse single molecule matrix to dense
}
