% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_calling.r
\name{MethSM.to.MethGR}
\alias{MethSM.to.MethGR}
\title{Compute MethGR from MethSM}
\usage{
MethSM.to.MethGR(MethSM, chromosome)
}
\arguments{
\item{MethSM}{internal CallContextMethylation}

\item{chromosome}{string}
}
\description{
Compute MethGR from MethSM
}
