% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs_utils.r
\name{panel.cor}
\alias{panel.cor}
\title{Utility for HighCoverage_MethRate_SampleCorrelation}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{digits}{number of digits}

\item{prefix}{string}

\item{cex.cor}{graphical param}
}
\description{
Utility for HighCoverage_MethRate_SampleCorrelation
}
