% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs_utils.r
\name{panel.hist}
\alias{panel.hist}
\title{Utility for HighCoverage_MethRate_SampleCorrelation}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{data for hist}

\item{...}{data for hist}
}
\description{
Utility for HighCoverage_MethRate_SampleCorrelation
}
