% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs_utils.r
\name{panel.jet}
\alias{panel.jet}
\title{Utility for HighCoverage_MethRate_SampleCorrelation}
\usage{
panel.jet(...)
}
\arguments{
\item{...}{data for lower pairs panel}
}
\description{
Utility for HighCoverage_MethRate_SampleCorrelation
}
