% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_plotting.r
\name{Plot_FootprintCharter_SM}
\alias{Plot_FootprintCharter_SM}
\title{Plot single molecule heatmaps of footprint detection results}
\usage{
Plot_FootprintCharter_SM(footprints.df, RegionOfInterest, partitions.order)
}
\arguments{
\item{RegionOfInterest}{GRanges interval to plot}

\item{partitions.order}{integer vector specifying the order in which to plot partitions}

\item{MethSM}{Single molecule matrix list as returned by CallContextMethylation()}
}
\description{
Plot single molecule heatmaps of footprint detection results
}
