% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs.r
\name{Plot_LowCoverageMethRate_RMSE}
\alias{Plot_LowCoverageMethRate_RMSE}
\title{Plot Low Coverage Methylation Rate RMSE}
\usage{
Plot_LowCoverageMethRate_RMSE(RMSE_DF)
}
\arguments{
\item{RMSE_DF}{data.frame as returned by the LowCoverageMethRate_RMSE function}
}
\description{
Produce barplot of RMSE values calculated for methylation rate distribution estimates of low coverage samples
}
