% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{PromoterStateQuantificationPlot}
\alias{PromoterStateQuantificationPlot}
\title{promoter state quantification bar}
\usage{
PromoterStateQuantificationPlot(SortedReads, states, Strand = "+")
}
\arguments{
\item{SortedReads}{List of sorted reads (can be multiple samples) as returned by SortReadsByTFCluster (or SortReads run with analogous parameters)}

\item{states}{as returned by TFpairStates function}

\item{Strand}{Set to either "+" or "-" when performing the promoter sorting, so that it will generate the plot with the correct strand}
}
\description{
promoter state quantification bar
}
