% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_calling_utils.r
\name{rbind.fill.Matrix}
\alias{rbind.fill.Matrix}
\title{Implementation performing a similar operation of plyr::rbind.fill.matrix but for sparseMatrix}
\usage{
\method{rbind}{fill.Matrix}(x, y)
}
\arguments{
\item{x}{sparse matrix constructed using the function Matrix::sparseMatrix. Should have Dimnames and dims (e.g. when indexing drop=FALSE)}

\item{y}{sparse matrix constructed using the function Matrix::sparseMatrix. Should have Dimnames and dims (e.g. when indexing drop=FALSE)}
}
\description{
Implementation performing a similar operation of plyr::rbind.fill.matrix but for sparseMatrix
}
\details{
N.b. only possible fill at the moment is 0
}
\examples{

Methylation = qs::qread(system.file("extdata", "Methylation_3.qs", 
package="SingleMoleculeFootprinting"))
MethSM_1 = Methylation[[2]][[1]]
Methylation = qs::qread(system.file("extdata", "Methylation_4.qs", 
package="SingleMoleculeFootprinting"))
MethSM_2 = Methylation[[2]][[1]]
rbind.fill.Matrix(MethSM_1, MethSM_2)

}
