% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_calling_utils.r
\name{rbind.fill.matrix.sparse}
\alias{rbind.fill.matrix.sparse}
\title{Implementation performing a similar operation of plyr::rbind.fill.matrix but for sparseMatrix}
\usage{
rbind.fill.matrix.sparse(x, y)
}
\arguments{
\item{x}{sparse matrix constructed using the function Matrix::sparseMatrix. Should have Dimnames and dims (e.g. when indexing drop=FALSE)}

\item{y}{sparse matrix constructed using the function Matrix::sparseMatrix. Should have Dimnames and dims (e.g. when indexing drop=FALSE)}
}
\description{
Implementation performing a similar operation of plyr::rbind.fill.matrix but for sparseMatrix
}
\details{
N.b. only possible fill at the moment is 0
}
\examples{

rbind.fill.matrix.sparse(MethSM_1, MethSM_2)

}
