% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_rolling_mean.r
\name{RollingMean}
\alias{RollingMean}
\title{Compute rolling mean}
\usage{
RollingMean(MethSM, RegionOfInterest, verbose = TRUE)
}
\arguments{
\item{MethSM}{sparse MethSM as returned by CallContextMethylation()}

\item{RegionOfInterest}{GRanges to analyse. Only the reads that cover continuously and entirely the range will be retained}

\item{verbose}{TRUE/FALSE}
}
\description{
higher level wrapper
}
\examples{

Methylation = qs::qread(system.file(
    "extdata", "Methylation_4.qs", package="SingleMoleculeFootprinting"
    ))
MethSM = Methylation[[2]]$SMF_MM_TKO_DE_
RegionOfInterest = GenomicRanges::GRanges("chr6", IRanges::IRanges(88106000, 88106500))
RegionOfInterest = IRanges::resize(RegionOfInterest, 80, "center")

}
