% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{rowMeans_drop0}
\alias{rowMeans_drop0}
\title{Calculate rowMeans after dropping zeros}
\usage{
rowMeans_drop0(MethSM)
}
\arguments{
\item{MethSM}{one single molecule sparse matrix}
}
\value{
rowMeans (N.b. this is +1 based)
}
\description{
Calculate rowMeans after dropping zeros
}
