% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{SingleTFStateQuantificationPlot}
\alias{SingleTFStateQuantificationPlot}
\title{Single TF state quantification bar}
\usage{
SingleTFStateQuantificationPlot(SortedReads)
}
\arguments{
\item{SortedReads}{Sorted reads as returned by SortReadsBySingleTF}
}
\description{
Single TF state quantification bar
}
