% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_sorting_utils.r
\name{SingleTFStates}
\alias{SingleTFStates}
\title{Hard-coded interpretation of biological states from single TF sorting}
\usage{
SingleTFStates()
}
\value{
list of states
}
\description{
Hard-coded interpretation of biological states from single TF sorting
}
\examples{

SingleTFStates()


}
