% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeWide_wrappers_to_SortingFunctions.R
\name{SortReadsByPromoter_MultiSiteWrapper}
\alias{SortReadsByPromoter_MultiSiteWrapper}
\title{Convenience wrapper to sort single molecule according to promoters at multiple TSSs in the genome}
\usage{
SortReadsByPromoter_MultiSiteWrapper(
  sampleSheet,
  sample,
  genome,
  coverage = 20,
  ConvRate.thr = 0.8,
  TSSsc,
  species = "MM",
  max_interTF_distance = 1e+05,
  max_window_width = 5e+06,
  min_cluster_width = 600,
  sorting_coverage = 30,
  cores = 1
)
}
\arguments{
\item{sampleSheet}{QuasR pointer file}

\item{sample}{samples to use, from the SampleName field of the sampleSheet}

\item{genome}{BSgenome}

\item{coverage}{coverage threshold as integer for least number of reads to cover a cytosine for it to be carried over in the analysis. Defaults to 20.}

\item{ConvRate.thr}{Convesion rate threshold. Double between 0 and 1, defaults to 0.8. To skip this filtering step, set to NULL. For more information, check out the details section.}

\item{TSSsc}{GRanges object of transcription factor binding sites coordinates}

\item{species}{species for promoter sorting (either "MM" for mouse or "DM" for Drosophila)}

\item{max_interTF_distance}{maximum distance between two consecutive TSSs for them to be grouped in the same window}

\item{max_window_width}{upper limit to window width. This value should be adjusted according to the user's system as it determines the amount of memory used in the later context methylation call}

\item{min_cluster_width}{lower limit to window width. Corresponds to the scenario when a window contains a single TSS}

\item{sorting_coverage}{integer. Minimum number of reads covering all sorting bins for sorting to be performed. Defaults to 30.}

\item{cores}{number of cores to use for parallel processing of multiple Methylation Calling Windows (i.e. groupings of adjecent TFBS clusters)}
}
\value{
list where [[1]] is the TSS GRanges object describing coordinates TSSs used to sort single molecules
                   [[2]] is a list of SortedReads nested per TSS and sample
                   [[3]] is a tibble reporting the count (and frequency) of reads per state, sample and TSS
}
\description{
The function starts from a list of single TSSs, arranges them into chuncks of nearby TSSs, calls methylation at the interested sites and outputs sorted reads
}
