% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{SortReadsBySinglePromoter}
\alias{SortReadsBySinglePromoter}
\title{Wrapper to SortReads for Promoter case}
\usage{
SortReadsBySinglePromoter(MethSM, TSS, coverage = 30)
}
\arguments{
\item{MethSM}{Single molecule matrix list as returned by CallContextMethylation}

\item{TSS}{Transcription Start Site to use for sorting, passed as a GRanges object of length >= 1}

\item{coverage}{integer. Minimum number of reads covering all sorting bins for sorting to be performed. Defaults to 30}

\item{species}{species for promoter sorting (either "MM" for mouse or "DM" for Drosophila)}
}
\value{
List of reads sorted by promoter
}
\description{
Wrapper to SortReads for Promoter case
}
\examples{

TSS = GenomicRanges::GRanges("chr1", IRanges(4785709), strand = "-")
elementMetadata(TSS)$gene_id = "NM_001177658.1"
names(TSS) = "NM_001177658.1"

SortedReads = SortReadsBySinglePromoter(MethSM = MethSM, TSS = TSS, species = "MM", coverage = 30)

}
