% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{SortReadsByTFCluster}
\alias{SortReadsByTFCluster}
\title{Wrapper to SortReads for TF cluster case}
\usage{
SortReadsByTFCluster(MethSM, TFBSs)
}
\arguments{
\item{MethSM}{Single molecule matrix}

\item{TFBSs}{Transcription factor binding sites to use for sorting, passed as a GRanges object of length > 1}
}
\value{
List of reads sorted by TF cluster
}
\description{
Wrapper to SortReads for TF cluster case
}
