% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{SortSingleMolecules}
\alias{SortSingleMolecules}
\title{Sort Single Molecules}
\usage{
SortSingleMolecules(
  regDF,
  sampleSheet,
  target_range,
  inMs = c(NA, NA),
  upMs = c(NA, NA),
  doMs = c(NA, NA),
  conv.rate = 0,
  remove.duplicates = F
)
}
\arguments{
\item{regDF}{Genomic chunk where single molecule methylation will be extracted (keep it ~1e8bp for optimal memory usage) provided as a data.frame}

\item{sampleSheet}{QuasR input file containing sample information (see \code{\link[QuasR]{qAlign}} documentation for more details). N.b. SampleNames have to be tagged with the type of treatment _NO_: only GCs will be used, _SS_: only CGs will be considered, _DE_: CG and GCs will be considered}

\item{target_range}{GRanges of TF binding motifs to be analyzed. N.b. If target range is a Grange, sorting will be carried in a single TF fashion. If is a GrangeList, reads will be sorted by TF clusters}

\item{inMs}{borders of the bin covering the TF binding site}

\item{doMs}{borders of the bin downstream of the binding site  (relative to the center of the motif)}

\item{conv.rate}{Integer from 0 to 100 representing the minimal accepted erroneous bisulfite conversion rate before discarding the read}

\item{remove.duplicates}{T/F. Remove identical reads, assumed to be PCR duplicates}

\item{upM}{borders of the bin upstream of the binding site (relative to the center of the motif)}
}
\description{
this
}
\details{
ugh
}
