% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{StateQuantificationPlot}
\alias{StateQuantificationPlot}
\title{Plot states quantification bar}
\usage{
StateQuantificationPlot(
  SortedReads,
  sorting.strategy = "classical",
  Strand = "+"
)
}
\arguments{
\item{SortedReads}{Sorted reads object as returned by SortReads function}

\item{sorting.strategy}{One of "classical" (default), "custom", "hierarchical.clustering" or "None".
Set to "classical" for classical one-TF/TF-pair sorting (as described in Sönmezer et al, MolCell, 2021). Should be passed along with argument SortedReads set to the Sorted reads object as returned by SortReads function.
If set to "custom", SortedReads should be a list with one item per sample (corresponding to MethSM).
If set to "hierarchical.clustering", the function will perform hierarchical clustering in place on a subset of reads. Useful to check for duplicated reads in amplicon sequencing experiments.
If set to "promoter", plots reads by promoter sorting
If set to "None", it will plot unsorted reads.}

\item{Strand}{Typically it is set to "+", but it has nothing to do though.
Set to either "+" or "-" when performing the promoter sorting, so that it will generate the plot with the correct strand}
}
\value{
Bar plot quantifying states
}
\description{
Plot states quantification bar
}
\examples{

Qinput = system.file("extdata", "QuasR_input_pairs.txt", package = "SingleMoleculeFootprinting", mustWork = T)
MySample = suppressMessages(readr::read_delim(Qinput, delim = "\t")[[2]])
Region_of_interest = GRanges(seqnames = "chr6", ranges = IRanges(start = 88106000, end = 88106500), strand = "*")
Methylation = CallContextMethylation(sampleSheet = Qinput,
                                     sample = MySample,
                                     genome = BSgenome.Mmusculus.UCSC.mm10,
                                     range = Region_of_interest,
                                     coverage = 20,
                                     ConvRate.thr = 0.2)
TFBSs = GenomicRanges::GRanges("chr6", IRanges(c(88106253), c(88106263)), strand = "-")
elementMetadata(TFBSs)$name = c("NRF1")
names(TFBSs) = c(paste0("TFBS_", c(4305216)))
SortedReads = SortReadsByTFCluster(MethSM = Methylation[[2]], TFBSs = TFBSs)

StateQuantificationPlot(SortedReads = SortedReads)

}
