% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs_utils.r
\name{SubsetGRangesForSamples}
\alias{SubsetGRangesForSamples}
\title{Subset Granges for given samples}
\usage{
SubsetGRangesForSamples(GRanges_obj, Samples)
}
\arguments{
\item{GRanges_obj}{GRanges object as returned by CallContextMethylation function}

\item{Samples}{vector of sample names as they appear in the SampleName field of the QuasR sampleFile}
}
\description{
Inner utility for LowCoverageMethRateDistribution
}
