% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{TFPairStateQuantificationPlot}
\alias{TFPairStateQuantificationPlot}
\title{TF pair state quantification bar}
\usage{
TFPairStateQuantificationPlot(SortedReads)
}
\arguments{
\item{SortedReads}{Sorted reads as returned by SortReadsByTFCluster}
}
\description{
TF pair state quantification bar
}
