% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSite.wrapper.R
\name{Unsupervised.clustering.MultiSite.wrapper}
\alias{Unsupervised.clustering.MultiSite.wrapper}
\title{This is useful for bait capture or genome-wide data when
wanting to cluster all covered sites before knowing which ones are covered}
\usage{
Unsupervised.clustering.MultiSite.wrapper(
  sampleSheet,
  Samples,
  pool.samples = TRUE,
  genome,
  TFBS.clusters,
  SNPs,
  MethylationCallingWindows,
  TFBSs,
  CytosinesToMask,
  species.string.match = "",
  deduplicate = FALSE,
  params = NULL,
  cores = 40,
  tryCatchLog.write.error.dump.file = FALSE
)
}
\arguments{
\item{pool.samples}{Whether to pool or not samples for clustering & footprints detection. Defaults to TRUE.}

\item{TFBS.clusters}{object resulting from a Arrange_TFBSs_clusters call}

\item{SNPs}{all SNPs (both species)}

\item{TFBSs}{Granges with all the annotated TFBSs}

\item{species.string.match}{Used to pick SNPs for plotting. Should be a match for how the Cast species is encoded in the SampleName string. Everything that doesn't match will be considered Spretus}

\item{deduplicate}{TRUE or FALSE (default). Useful with samples with high PCR duplication}

\item{params}{bait.capture.parameters.list}

\item{cores}{}
}
\description{
This is useful for bait capture or genome-wide data when
wanting to cluster all covered sites before knowing which ones are covered
}
