% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs.r
\name{BaitCapture}
\alias{BaitCapture}
\title{Bait capture efficiency}
\usage{
BaitCapture(sampleSheet, genome, paired = "fr", baits, clObj = NULL)
}
\arguments{
\item{sampleSheet}{QuasR sample sheet}

\item{genome}{BS genome}

\item{paired}{paired type of paired end library. Defaults to "fr". Refer to QuasR::qAlign documentation for more details}

\item{baits}{GRanges obj of bait coordinates. We provide and example through SingleMoleculeFootprintingData::EnrichmentRegions_mm10.rds()}

\item{clObj}{cluster object to emply for parallel processing created using the parallel::makeCluster function. Defaults to NULL}
}
\value{
bait capture efficiency
}
\description{
check bait capture efficiency. Expected to be ~70% for mouse genome
}
\examples{
Qinput = paste0(tempdir(), "/NRF1Pair_Qinput.txt")
library(BSgenome.Mmusculus.UCSC.mm10)

if(file.exists(Qinput)){
    # DO NOT RUN
    # clObj = parallel::makeCluster(5)
    # BaitRegions = SingleMoleculeFootprintingData::EnrichmentRegions_mm10.rds()
    # BaitCaptureEfficiency = BaitCapture(sampleSheet = Qinput, genome = BSgenome.Mmusculus.UCSC.mm10, baits = BaitRegions, clObj=clObj)
    # parallel::stopCluster(clObj)
}

}
