% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{BinMethylation}
\alias{BinMethylation}
\title{Summarize methylation inside sorting bins}
\usage{
BinMethylation(MethSM, TFBS, bin)
}
\arguments{
\item{MethSM}{Single molecule matrix}

\item{TFBS}{Transcription factor binding site to use for sorting, passed as a GRanges object of length 1}

\item{bin}{vector of two integers representing the coordinate of a bin relative to the center of the TFBS}
}
\value{
Reads covering bin with their summarized methylation status
}
\description{
Summarize methylation inside sorting bins
}
