% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{CallContextMethylation}
\alias{CallContextMethylation}
\title{Call Context Methylation}
\usage{
CallContextMethylation(meth_gr, cO, genome)
}
\arguments{
\item{meth_gr}{Methylation GRange object as returned by QuasR function qMeth}

\item{genome}{Primary sequence of genome of interest. E.g. for BSgenome.Mmusculus.UCSC.mm10 this would be Mmusculus}

\item{c0}{coverage threshold}
}
\description{
Call Context Methylation
}
