% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{CallContextMethylation}
\alias{CallContextMethylation}
\title{Call Context Methylation}
\usage{
CallContextMethylation(
  sampleSheet,
  sample,
  genome,
  paired = "fr",
  RegionOfInterest,
  coverage = 20,
  ConvRate.thr = 0.8,
  returnSM = TRUE,
  clObj = NULL
)
}
\arguments{
\item{sampleSheet}{QuasR pointer file}

\item{sample}{for now this works for sure on one sample at the time only}

\item{genome}{BSgenome}

\item{paired}{paired type of paired end library. Defaults to "fr". Refer to QuasR::qAlign documentation for more details}

\item{RegionOfInterest}{GenimocRange representing the genomic region of interest}

\item{coverage}{coverage threshold as integer for least number of reads to cover a cytosine for it to be carried over in the analysis. Defaults to 20.}

\item{ConvRate.thr}{Convesion rate threshold. Double between 0 and 1, defaults to 0.8. To skip this filtering step, set to NULL. For more information, check out the details section.}

\item{returnSM}{whether to return the single molecule matrix, defaults to TRUE}

\item{clObj}{cluster object for parallel processing of multiple samples. For now only used by qMeth call for bulk methylation. Should be the output of a parallel::makeCluster() call}
}
\value{
List with two Granges objects: average methylation call (GRanges) and single molecule methylation call (matrix)
}
\description{
Can deal with multiple samples
}
\details{
The ConvRate.thr argument should be used with care as it could create biases (e.g. when only one C out of context is present) while generally only marginally cleaning up the data.
}
\examples{

Qinput = system.file("extdata", "QuasR_input_pairs.txt", package = "SingleMoleculeFootprinting", mustWork = T)
MySample = suppressMessages(readr::read_delim(Qinput, delim = "\t")[[2]])
Region_of_interest = GRanges(seqnames = "chr6", ranges = IRanges(start = 88106000, end = 88106500), strand = "*")

Methylation = CallContextMethylation(sampleSheet = Qinput,
                                     sample = MySample,
                                     genome = BSgenome.Mmusculus.UCSC.mm10,
                                     RegionOfInterest = Region_of_interest,
                                     coverage = 20,
                                     ConvRate.thr = 0.2)

}
