% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompositePlotting.R
\name{CollectCompositeData}
\alias{CollectCompositeData}
\title{Collect bulk SMF data for later composite plotting}
\usage{
CollectCompositeData(
  sampleSheet,
  sample,
  genome,
  TFBSs,
  window,
  coverage = 20,
  ConvRate.thr = 0.8,
  minCytosines = 0,
  cores = 1
)
}
\arguments{
\item{sampleSheet}{QuasR pointer file}

\item{sample}{for now this works for sure on one sample at the time only}

\item{genome}{BSgenome}

\item{TFBSs}{GRanges object of TF binding sites to collect info for. We reccommend employing 50 to 200 TFBSs.}

\item{window}{window size to collect methylation information for}

\item{coverage}{coverage threshold as integer for least number of reads to cover a cytosine for it to be carried over in the analysis. Defaults to 20.}

\item{ConvRate.thr}{Convesion rate threshold. Double between 0 and 1, defaults to 0.8. For more information, check out the details section}

\item{minCytosines}{minimum number of cytosines covered per TFBS}

\item{cores}{number of cores to use}
}
\value{
data.frame of bulk SMF info ready for plotting
}
\description{
Collect bulk SMF data for later composite plotting
}
