% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompositePlotting.R
\name{CollectCompositeData}
\alias{CollectCompositeData}
\title{Collect bulk SMF data for later composite plotting}
\usage{
CollectCompositeData(
  sampleSheet,
  sample,
  genome,
  TFBSs,
  minCytosines = 0,
  cores = 20
)
}
\arguments{
\item{sampleSheet}{QuasR pointer file}

\item{sample}{for now this works for sure on one sample at the time only}

\item{genome}{BSgenome}

\item{TFBSs}{GRanges object of TF binding sites to collect info for}

\item{minCytosines}{minimum number of cytosines covered per TFBS}

\item{cores}{number of cores}
}
\value{
data.frame of bulk SMF info ready for plotting
}
\description{
Collect bulk SMF data for later composite plotting
}
