% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs.r
\name{ConversionRate}
\alias{ConversionRate}
\title{Conversion rate}
\usage{
ConversionRate(sampleSheet, genome, paired = "fr", chr = 19, cores = 1)
}
\arguments{
\item{sampleSheet}{QuasR sample sheet}

\item{genome}{BS genome}

\item{paired}{paired type of paired end library. Defaults to "fr". Refer to QuasR::qAlign documentation for more details}

\item{chr}{chromosome to calculate conversion rate on (default: 19)}

\item{cores}{number of cores for parallel processing. Defaults to 1}
}
\description{
calculate sequencing library conversion rate on a chromosome of choice
}
\examples{

Qinput = system.file("extdata", "QuasR_input_pairs.txt", package = "SingleMoleculeFootprinting", mustWork = T)

ConversionRatePrecision = ConversionRate(sampleSheet = Qinput, genome = BSgenome.Mmusculus.UCSC.mm10, chr = 19, cores = 1)

}
