% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_QCs.r
\name{ConversionRate}
\alias{ConversionRate}
\title{Conversion rate}
\usage{
ConversionRate(sampleSheet, genome, chr = 19)
}
\arguments{
\item{sampleSheet}{QuasR sample sheet}

\item{genome}{BS genome}

\item{chr}{chromosome to calculate conversion rate on (default: 19)}
}
\description{
calculate sequencing library conversion rate on a chromosome of choice
}
