% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeWide_wrappers_to_SortingFunctions.R
\name{Create_MethylationCallingWindows}
\alias{Create_MethylationCallingWindows}
\title{Create methylation calling windows to call context methylation in one run for clusters lying proximally to each other}
\usage{
Create_MethylationCallingWindows(
  TFBS_cluster_coordinates,
  max_intercluster_distance = 1e+05,
  max_window_width = 5e+06,
  min_cluster_width = 600
)
}
\arguments{
\item{TFBS_cluster_coordinates}{TFBS cluster coordinates analogous to ClusterCoordinates object returned by Arrange_TFBSs_clusters function}

\item{max_intercluster_distance}{maximum distance between two consecutive TFBS clusters for them to be grouped in the same window}

\item{max_window_width}{upper limit to window width. This value should be adjusted according to the user's system as it determines the amount of memory used in the later context methylation call}

\item{min_cluster_width}{lower limit to window width. Corresponds to the scenario when a window contains a single TFBS cluster.}
}
\value{
GRanges object of window coordinates to be used for more efficient calls of CallContextMethylation
}
\description{
Relevant for genome-wide analyses
}
