% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{DetectExperimentType}
\alias{DetectExperimentType}
\title{Detect type of experiment}
\usage{
DetectExperimentType(Samples)
}
\arguments{
\item{Samples}{SampleNames field from QuasR sampleSheet}
}
\description{
Detect type of experiment
}
