% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{DetectExperimentType}
\alias{DetectExperimentType}
\title{Detect type of experiment}
\usage{
DetectExperimentType(Samples, verbose = TRUE)
}
\arguments{
\item{Samples}{SampleNames field from QuasR sampleSheet}

\item{verbose}{TRUE/FALSE}
}
\value{
String indicating the type of experiment detected
}
\description{
Detect type of experiment
}
\examples{
Qinput = paste0(tempdir(), "/NRF1Pair_Qinput.txt")

if(file.exists(Qinput)){
    sample = readr::read_delim(Qinput, delim = "\t")$SampleName
    ExpType = DetectExperimentType(sample)
}

}
