% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_rolling_mean.r
\name{fill.empty.columns}
\alias{fill.empty.columns}
\title{Fills empty columns}
\usage{
fill.empty.columns(MethSM, verbose = TRUE)
}
\arguments{
\item{MethSM}{coming from a matrix.sliding.window.average call}

\item{verbose}{TRUE/FALSE}
}
\description{
when computing the sliding window matrix sometimes there are columns that are entirely NAs
because two nearest cytosines are further away from each other than the size of the sliding window used.
As a solution we fill these columns with the mean value from the two surrounding cytosines
(as long as the number of columns to be entirely NAs is lower than 20)
}
