% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_rolling_mean.r
\name{fill.empty.columns}
\alias{fill.empty.columns}
\title{Fills empty columns}
\usage{
\method{fill}{empty.columns}(MethSM, verbose = TRUE)
}
\arguments{
\item{MethSM}{coming from a matrix.sliding.window.average call}

\item{verbose}{TRUE/FALSE}
}
\description{
PROBLEM: when computing the sliding window matrix sometimes there are colummns that are etnirely NAs
         because two nearest cytosines are further away from each other than the size of the sliding window used.
         As a ducktape solution to keep working I'll fill these columns with the mean value from the two surorunding cytosines
         (as long as the number of columns to be completely NAs is lower than 20)
}
