% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_rolling_mean.r
\name{filter.dense.matrix}
\alias{filter.dense.matrix}
\title{Filters dense matrix}
\usage{
\method{filter}{dense.matrix}(MethSM, RegionOfInterest, verbose = TRUE)
}
\arguments{
\item{RegionOfInterest}{GRanges to analyse. Only the reads that cover continuously and entirely the range will be retained}

\item{verbose}{TRUE/FALSE}
}
\description{
Filters dense matrix
}
