% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{FilterByConversionRate}
\alias{FilterByConversionRate}
\title{Calculate reads conversion rate}
\usage{
FilterByConversionRate(MethSM, chr, genome, thr)
}
\arguments{
\item{MethSM}{as comes out of the func GetSingleMolMethMat}

\item{chr}{Chromosome, MethSM doesn't carry this info}

\item{genome}{BSgenome}

\item{thr}{Double between 0 and 1. Threshold below which to filter reads.}
}
\value{
Filtered MethSM
}
\description{
Calculate reads conversion rate
}
\examples{

Qinput = system.file("extdata", "QuasR_input_pairs.txt", package = "SingleMoleculeFootprinting", mustWork = T)
QuasRprj = GetQuasRprj(Qinput, BSgenome.Mmusculus.UCSC.mm10)
sample = suppressMessages(readr::read_delim(Qinput, delim = "\t")[[2]])
range = GRanges(seqnames = "chr6", ranges = IRanges(start = 88106000, end = 88106500), strand = "*")
MethSM = GetSingleMolMethMat(QuasRprj, range, sample)

MethSM = FilterByConversionRate(MethSM, chr = "chr19", genome = BSgenome.Mmusculus.UCSC.mm10, thr = 0.8)

}
