% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{FilterByConversionRate}
\alias{FilterByConversionRate}
\title{Calculate reads conversion rate}
\usage{
FilterByConversionRate(MethSM, chr, genome, thr = 0.2)
}
\arguments{
\item{MethSM}{as comes out of the func GetSingleMolMethMat}

\item{chr}{Chromosome, MethSM doesn't carry this info}

\item{genome}{BSgenome}

\item{thr}{Double between 0 and 1. Threshold above which to filter reads. Defaults to 0.2}
}
\value{
Filtered MethSM
}
\description{
Calculate reads conversion rate
}
