% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{FilterContextCytosines}
\alias{FilterContextCytosines}
\title{Filter Cytosines in context}
\usage{
FilterContextCytosines(MethGR, genome, context)
}
\arguments{
\item{MethGR}{Granges obj of average methylation}

\item{genome}{BSgenome}

\item{context}{Context of interest (e.g. "GC", "CG",..)}
}
\value{
filtered Granges obj
}
\description{
Filter Cytosines in context
}
\examples{
Qinput = paste0(tempdir(), "/NRF1Pair_Qinput.txt")
library(BSgenome.Mmusculus.UCSC.mm10)

if(file.exists(Qinput)){
    QuasRprj = GetQuasRprj(Qinput, BSgenome.Mmusculus.UCSC.mm10)

    Samples = readr::read_delim(Qinput, delim = "\t")$SampleName
    sample = Samples[1]
    range = GRanges(seqnames = "chr6", ranges = IRanges(start = 88106000, end = 88106500), strand = "*")

    MethGR = QuasR::qMeth(QuasRprj[grep(sample, Samples)], mode="allC", range, collapseBySample = TRUE, keepZero = TRUE)
    FilterContextCytosines(MethGR, BSgenome.Mmusculus.UCSC.mm10, "NGCNN")
}

}
