% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{GetQuasRprj}
\alias{GetQuasRprj}
\title{Get QuasRprj}
\usage{
GetQuasRprj(sampleSheet, genome)
}
\arguments{
\item{sampleSheet}{QuasR pointer file}

\item{genome}{BSgenome}
}
\value{
QuasR project object as returned by QuasR::qAlign function
}
\description{
Get QuasRprj
}
\examples{

Qinput = system.file("extdata", "QuasR_input_pairs.txt", package = "SingleMoleculeFootprinting", mustWork = TRUE)
QuasRprj = GetQuasRprj(Qinput, BSgenome.Mmusculus.UCSC.mm10)

}
