% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{GetQuasRprj}
\alias{GetQuasRprj}
\title{Get QuasRprj}
\usage{
GetQuasRprj(sampleSheet, genome)
}
\arguments{
\item{sampleSheet}{QuasR pointer file}

\item{genome}{BSgenome}
}
\value{
QuasR project object as returned by QuasR::qAlign function
}
\description{
Get QuasRprj
}
\examples{
Qinput = paste0(tempdir(), "/NRF1Pair_Qinput.txt")
library(BSgenome.Mmusculus.UCSC.mm10)

if(file.exists(Qinput)){
    QuasRprj = GetQuasRprj(Qinput, BSgenome.Mmusculus.UCSC.mm10)
}

}
