% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{HierarchicalClustering}
\alias{HierarchicalClustering}
\title{Perform Hierarchical clustering on single reads}
\usage{
HierarchicalClustering(MethSM)
}
\arguments{
\item{MethSM}{Single molecule methylation matrix}
}
\description{
Perform Hierarchical clustering on single reads
}
