% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_methylation_functions.r
\name{MethSM.to.MethGR}
\alias{MethSM.to.MethGR}
\title{Compute MethGR from MethSM}
\usage{
MethSM.to.MethGR(MethSM, chromosome)
}
\arguments{
\item{chromosome}{}
}
\description{
Compute MethGR from MethSM
}
