% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{PlotAvgSMF}
\alias{PlotAvgSMF}
\title{Plot average methylation}
\usage{
PlotAvgSMF(MethGR, range, TFBSs)
}
\arguments{
\item{MethGR}{Average methylation GRanges obj}

\item{range}{GRanges interval to plot}

\item{TFBSs}{GRanges object of transcription factor binding sites to include in the plot. Assumed to be already subset.}
}
\description{
Plot average methylation
}
