% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintCharter_plotting.r
\name{PlotFootprints}
\alias{PlotFootprints}
\title{Plot bulk SMF separately for each partition alognside footprint detection results.}
\usage{
PlotFootprints(MethSM, partitioned.molecules, footprints.df, TFBSs)
}
\arguments{
\item{MethSM}{sparse MethSM as returned by CallContextMethylation()}

\item{partitioned.molecules}{vector of partition assignments per molecule as returned by FootprintCharter()}

\item{footprints.df}{data.frame of footprints as returned by FootprintCharter()}

\item{TFBSs}{TFBSs annotation at the RegionOfInterest (optional).}
}
\description{
Plot bulk SMF separately for each partition alognside footprint detection results.
}
