% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{PlotSingleMoleculeStack}
\alias{PlotSingleMoleculeStack}
\title{Plot single molecule stack}
\usage{
PlotSingleMoleculeStack(MethSM, range)
}
\arguments{
\item{MethSM}{Single molecule methylation matrix}

\item{range}{GRanges interval to plot}
}
\value{
Single molecule plot
}
\description{
Plot single molecule stack
}
