% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{PlotSingleMoleculeStack}
\alias{PlotSingleMoleculeStack}
\title{Plot single molecule stack}
\usage{
PlotSingleMoleculeStack(MethSM, RegionOfInterest)
}
\arguments{
\item{MethSM}{Single molecule methylation matrix}

\item{RegionOfInterest}{GRanges interval to plot}
}
\description{
Plot single molecule stack
}
\examples{

Qinput = system.file("extdata", "QuasR_input_pairs.txt", package = "SingleMoleculeFootprinting", mustWork = T)
MySample = suppressMessages(readr::read_delim(Qinput, delim = "\t")[[2]])
Region_of_interest = GRanges(seqnames = "chr6", ranges = IRanges(start = 88106000, end = 88106500), strand = "*")
Methylation = CallContextMethylation(sampleSheet = Qinput,
                                     sample = MySample,
                                     genome = BSgenome.Mmusculus.UCSC.mm10,
                                     range = Region_of_interest,
                                     coverage = 20,
                                     ConvRate.thr = 0.2)

PlotSingleMoleculeStack(MethSM = Methylation[[2]], RegionOfInterest = Region_of_interest)

}
