% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{PlotSingleSiteSMF}
\alias{PlotSingleSiteSMF}
\title{Plot SMF data at single site}
\usage{
PlotSingleSiteSMF(
  ContextMethylation,
  sample,
  range,
  SortedReads = NULL,
  TFBSs,
  saveAs = NULL
)
}
\arguments{
\item{ContextMethylation}{Context methylation object as returned by CallContextMethylation function}

\item{sample}{one sample as reported in the SampleName files of the QuasR sampleSheet}

\item{range}{GRange interval to plot}

\item{SortedReads}{Defaults to NULL, in which case will plot unsorted reads. Sorted reads object as returned by SortReads function or "HC" to perform hierarchical clustering}

\item{TFBSs}{GRange or GRangesList of transcription factor binding sites to add to the plot. If SortedReads are passed, the format of TFBSs (GRanges vs GRangesList) will be used to determie if single molecules were sorted based on one or multiple TFs}

\item{saveAs}{Full path to pdf file to save plot to. Defaults to NULL, in which case will only display}
}
\description{
Plot SMF data at single site
}
