% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{PlotSM}
\alias{PlotSM}
\title{Wrapper for PlotSingleMoleculeStack function}
\usage{
PlotSM(MethSM, range, SortedReads = NULL)
}
\arguments{
\item{MethSM}{Single molecule methylation matrix}

\item{range}{GRanges interval to plot}

\item{SortedReads}{Defaults to NULL, in which case will plot unsorted reads. Sorted reads object as returned by SortReads function or "HC" to perform hierarchical clustering}
}
\description{
adds the convenience of arranging reads before plotting
}
