% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{SingleTFStateQuantificationPlot}
\alias{SingleTFStateQuantificationPlot}
\title{Single TF state quantification bar}
\usage{
SingleTFStateQuantificationPlot(states, OrderedReads)
}
\arguments{
\item{states}{as returned by OneTFstates function}

\item{OrderedReads}{Reads ordered by states}
}
\description{
Single TF state quantification bar
}
