% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{SingleTFStateQuantificationPlot}
\alias{SingleTFStateQuantificationPlot}
\title{Single TF state quantification bar}
\usage{
SingleTFStateQuantificationPlot(SortedReads, states)
}
\arguments{
\item{SortedReads}{List of sorted reads (can be multiple samples) as returned by SortReadsBySingleTF (or SortReads run with analogous parameters)}

\item{states}{as returned by OneTFstates function}
}
\description{
Single TF state quantification bar
}
