% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{SortReads}
\alias{SortReads}
\title{Sort reads by single TF}
\usage{
SortReads(MethSM, BinsCoordinates, coverage = NULL)
}
\arguments{
\item{MethSM}{Single molecule matrix}

\item{BinsCoordinates}{IRanges object of absolute coordinates for sorting bins}

\item{coverage}{integer. Minimum number of reads covering all sorting bins for sorting to be performed}
}
\value{
list of sorted reads
}
\description{
Sort reads by single TF
}
\examples{

BinsCoord = list(c(-35,-25), c(-15,15), c(25,35))
TFBSs = GenomicRanges::GRanges("chr6", IRanges(c(88106253), c(88106263)), strand = "-")
elementMetadata(TFBSs)$name = c("NRF1")
names(TFBSs) = c(paste0("TFBS_", c(4305216)))

TFBS_center = start(TFBS) + (end(TFBS)-start(TFBS))/2
BinsCoordinates = IRanges(start = c(TFBS_center+bins[[1]][1], TFBS_center+bins[[2]][1], TFBS_center+bins[[3]][1]),
                          end = c(TFBS_center+bins[[1]][2], TFBS_center+bins[[2]][2], TFBS_center+bins[[3]][2]))

SortedReads = SortReads(MethSM, BinsCoordinates)

}
