% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{SortReads}
\alias{SortReads}
\title{Sort reads by single TF}
\usage{
SortReads(MethSM, TFBS, BinsCoord, SortByCluster)
}
\arguments{
\item{MethSM}{Single molecule matrix}

\item{TFBS}{Transcription factor binding site to use for sorting}

\item{BinsCoord}{list of 3 bin coordinates relative to the center of the TFBS.}

\item{SortByCluster}{T/F}
}
\value{
list of sorted reads
}
\description{
Sort reads by single TF
}
