% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{SortReadsBySingleTF}
\alias{SortReadsBySingleTF}
\title{Wrapper to SortReads for single TF case}
\usage{
SortReadsBySingleTF(
  MethSM,
  TFBS,
  bins = list(c(-35, -25), c(-15, 15), c(25, 35))
)
}
\arguments{
\item{MethSM}{Single molecule matrix list as returned by CallContextMethylation}

\item{TFBS}{Transcription factor binding site to use for sorting, passed as a GRanges object of length 1}

\item{bins}{list of 3 relative bin coordinates. Defaults to list(c(-35,-25), c(-15,15), c(25,35)).
bins[[1]] represents the upstream bin, with coordinates relative to the start of the TFBS.
bins[[2]] represents the TFBS bin, with coordinates relative to the center of the TFBS.
bins[[3]] represents the downstream bin, with coordinates relative to the end of the TFBS.}
}
\value{
List of reads sorted by single TF
}
\description{
Wrapper to SortReads for single TF case
}
\examples{

TFBSs = GenomicRanges::GRanges("chr6", IRanges(c(88106253), c(88106263)), strand = "-")
elementMetadata(TFBSs)$name = c("NRF1")
names(TFBSs) = c(paste0("TFBS_", c(4305216)))

SortedReads = SortReadsBySingleTF(MethSM = MethSM, TFBS = TFBSs)

}
