% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{SortReadsByTFCluster}
\alias{SortReadsByTFCluster}
\title{Wrapper to SortReads for TF cluster case}
\usage{
SortReadsByTFCluster(MethSM, TFBSs)
}
\arguments{
\item{MethSM}{Single molecule matrix}

\item{TFBSs}{Transcription factor binding sites to use for sorting, passed as a GRanges object of length > 1}
}
\value{
List of reads sorted by TF cluster
}
\description{
Wrapper to SortReads for TF cluster case
}
\examples{

TFBSs = GenomicRanges::GRanges("chr6", IRanges(c(88106216, 88106253), c(88106226, 88106263)), strand = "-")
elementMetadata(TFBSs)$name = c("NRF1", "NRF1")
names(TFBSs) = c(paste0("TFBS_", c(4305215, 4305216)))

SortedReads = SortReadsByTFCluster(MethSM = MethSM, TFBSs = TFBSs)

}
