% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{SortReadsByTFCluster}
\alias{SortReadsByTFCluster}
\title{Wrapper to SortReads for TF cluster case}
\usage{
SortReadsByTFCluster(
  MethSM,
  TFBS_cluster,
  bins = list(c(-35, -25), c(-7, 7), c(25, 35))
)
}
\arguments{
\item{MethSM}{Single molecule matrix}

\item{TFBS_cluster}{Transcription factor binding sites to use for sorting, passed as a GRanges object of length > 1}

\item{bins}{list of 3 relative bin coordinates. Defaults to list(c(-35,-25), c(-7,7), c(25,35)).
bins[[1]] represents the upstream bin, with coordinates relative to the start of the most upstream TFBS.
bins[[2]] represents all the TFBS bins, with coordinates relative to the center of each TFBS.
bins[[3]] represents the downstream bin, with coordinates relative to the end of the most downstream TFBS.}
}
\value{
List of reads sorted by TF cluster
}
\description{
Wrapper to SortReads for TF cluster case
}
\examples{

TFBSs = GenomicRanges::GRanges("chr6", IRanges(c(88106216, 88106253), c(88106226, 88106263)), strand = "-")
elementMetadata(TFBSs)$name = c("NRF1", "NRF1")
names(TFBSs) = c(paste0("TFBS_", c(4305215, 4305216)))

SortedReads = SortReadsByTFCluster(MethSM = MethSM, TFBSs = TFBS_cluster)

}
