% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{StateQuantification}
\alias{StateQuantification}
\title{Convenience for calculating state frequencies}
\usage{
StateQuantification(SortedReads, states)
}
\arguments{
\item{SortedReads}{List of sorted reads (can be multiple samples) as returned by either read sorting function (SortReads, SortReadsBySingleTF, SortReadsByTFCluster)}

\item{states}{states reporting the biological interpretation of patterns as return by either OneTFstates or TFpairStates functions.}
}
\value{
tibble with state frequency information
}
\description{
Convenience for calculating state frequencies
}
