% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_molecule_manipulation_functions.r
\name{StateQuantificationByTFPair}
\alias{StateQuantificationByTFPair}
\title{Convenience for calculating state frequencies after sorting reads by TF pair}
\usage{
StateQuantificationByTFPair(SortedReads)
}
\arguments{
\item{SortedReads}{List of sorted reads (can be multiple samples) as returned by SortReadsByTFCluster run for clusters of size 2 (or SortReads run with analogous parameters)}
}
\value{
tibble with state frequency information
}
\description{
wraps around StateQuantification function
}
