% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{StateQuantificationPlot}
\alias{StateQuantificationPlot}
\title{Plot states quantification bar}
\usage{
StateQuantificationPlot(SortedReads)
}
\arguments{
\item{SortedReads}{Sorted reads object as returned by SortReads function}
}
\value{
Bar plot quantifying states
}
\description{
Plot states quantification bar
}
