% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{StateQuantificationPlot}
\alias{StateQuantificationPlot}
\title{Plot states quantification bar}
\usage{
StateQuantificationPlot(SortedReads)
}
\arguments{
\item{SortedReads}{Sorted reads object as returned by SortReads function}
}
\value{
Bar plot quantifying states
}
\description{
Plot states quantification bar
}
\examples{

Qinput = system.file("extdata", "QuasR_input_pairs.txt", package = "SingleMoleculeFootprinting", mustWork = T)
MySample = suppressMessages(readr::read_delim(Qinput, delim = "\t")[[2]])
Region_of_interest = GRanges(seqnames = "chr6", ranges = IRanges(start = 88106000, end = 88106500), strand = "*")
Methylation = CallContextMethylation(sampleSheet = Qinput,
                                     sample = MySample,
                                     genome = BSgenome.Mmusculus.UCSC.mm10,
                                     range = Region_of_interest,
                                     coverage = 20,
                                     ConvRate.thr = 0.2)
TFBSs = GenomicRanges::GRanges("chr6", IRanges(c(88106253), c(88106263)), strand = "-")
elementMetadata(TFBSs)$name = c("NRF1")
names(TFBSs) = c(paste0("TFBS_", c(4305216)))
SortedReads = SortReadsByTFCluster(MethSM = Methylation[[2]], TFBSs = TFBSs)

StateQuantificationPlot(SortedReads = SortedReads)

}
