% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{TFPairStateQuantificationPlot}
\alias{TFPairStateQuantificationPlot}
\title{TF pair state quantification bar}
\usage{
TFPairStateQuantificationPlot(SortedReads, states)
}
\arguments{
\item{SortedReads}{List of sorted reads (can be multiple samples) as returned by SortReadsByTFCluster (or SortReads run with analogous parameters)}

\item{states}{as returned by TFpairStates function}
}
\description{
TF pair state quantification bar
}
