% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_molecule_examples.r
\name{TFPairStateQuantificationPlot}
\alias{TFPairStateQuantificationPlot}
\title{TF pair state quantification bar}
\usage{
TFPairStateQuantificationPlot(states, OrderedReads)
}
\arguments{
\item{states}{as returned by TFpairStates function}

\item{OrderedReads}{Reads ordered by states}
}
\value{
TF pair state quantification plot
}
\description{
TF pair state quantification bar
}
